Importing the 2017bot vision code as a starting point.
We can't seem to find a C++ NetworkTables library.  If that can't be
found, we will use the known and tested socket-based client- and
server code from last year (with the caveat that we need to be careful
about the hostname of the target RoboRIO.)

The other ting we'll need to master is the HttpCamera class.  That,
fortunately, is part of the C++ WPILib network interface.